(CTS)(Breaking) Prototype CTS texts have been moved
From MyCapytain.resources.prototypes.text to MyCapytain.resources.prototypes.cts.text :

- CtsNode now PrototypeCtsNode
- CtsPassage now PrototypeCtsPassage
- CtsText now PrototypeCtsText
